Split permission slips and resources into separate pages
- Created new PermissionSlipDashboard class in permission_slip_dashboard.js
- Updated ResourceDashboard to remove permission slip functionality
- Modified router.js to handle /permission-slips and /resources as separate routes
- Added home button to both pages for easy navigation back to dashboard
- Enhanced mobile-first responsive design:
  - Added summary-grid, summary-tile, and summary-list CSS classes
  - Added responsive breakpoint for .grid-2 to stack on mobile (max-width: 768px)
  - Ensured tables are horizontally scrollable on mobile devices
- Both pages now follow UX best practices for mobile usage
