+++ css/styles.css
@@ -319,6 +319,50 @@ a:focus-visible {
   grid-template-columns: repeat(auto-fit, minmax(min(100%, 250px), 1fr));
 }
 
+/* Summary Dashboard Grids */
+.summary-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(min(100%, 280px), 1fr));
+  gap: var(--space-md);
+  margin-block-end: var(--space-md);
+}
+
+.summary-tile {
+  background-color: var(--color-surface);
+  border-radius: var(--radius-md);
+  padding: var(--space-md);
+  box-shadow: var(--shadow-sm);
+  border: 1px solid var(--color-border-light);
+}
+
+.summary-label {
+  font-weight: var(--font-weight-semibold);
+  color: var(--color-primary);
+  margin-block-end: var(--space-sm);
+  font-size: var(--font-size-lg);
+}
+
+.summary-list {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.summary-list li {
+  padding: var(--space-xs) 0;
+  border-bottom: 1px solid var(--color-border-light);
+  font-size: var(--font-size-base);
+}
+
+.summary-list li:last-child {
+  border-bottom: none;
+}
+
+.summary-list strong {
+  color: var(--color-primary-dark);
+  font-weight: var(--font-weight-bold);
+}
+
 /* ============================================================================
    5. COMPONENT LIBRARY
    ============================================================================ */
@@ -6310,6 +6354,11 @@ details p {
     display: block;
     overflow-x: auto;
     white-space: nowrap;
+  }
+
+  /* Stack grid-2 layout on mobile for better readability */
+  .grid-2 {
+    grid-template-columns: 1fr;
   }
 }
